Fix use of wrong argument to EndOfTag
The beginning of a tag is searched for in 'string', so we have to search
for the end of the tag in 'string' as well, not 'token'. Otherwise we
fail to find the end and will emit uncompilable code in case there is
more than one @param tag for a function.